Level Assets Update and Camera Update
Floor and Wall assets were given mesh components. Camera updated to use Cinemachine with a Virtual Camera to follow player. Target button was created to be used to manually reset camera to face the player's look direction and controls were updated to be camera relative instead of world relative for smoother turning and control.